requirements.txt
Flask==1.1.1
gunicorn==19.9.0
line-bot-sdk==1.18.0
